refactor: 메시지 공통 기능 리팩토링
- 당첨 내역 메시지 추가 시 사용될 공통 기능을 추출할 필요성 발생.
- 블록 사이에 구분자 처리, 인라인 사이에 구분자 처리 등 공통 기능을 Formatter로 구분.
- 발권 정보 문자열 기능은 IssueMessage로 이전.
- 기존의 FormatterTest를 MessageTest로 이름을 변경해, 테스트 의미의 명확성을 높임.
