+++ src/main/java/lotto/messge/Formatter.java
@@ -1,0 +1,14 @@
+package lotto.messge;
+
+import java.util.List;
+
+public class Formatter {
+
+    public static String blocks(String... blocks) {
+        return String.join("\n", blocks);
+    }
+
+    public static String inlines(List<String> inlines) {
+        return String.join("\n", inlines);
+    }
+}
